LPS-83949 Move resolution logic to its own package
